Armado final de relaciones entre Eventos y Promos
-armado de relacion a mano.
-la idea es migrar a los metodos de mongoose

+++ services/index.js
@@ -65,7 +65,8 @@ function checkValidity(array) {
 }
 
 function updateFile(updateId, fileNew, fileLast){
-    if (fileLast) {
+
+    if (fileLast && fileLast !== 'delete') {
       if (fileNew !== fileLast) {
         fs.unlink(fileLast, (err) => {
           if (err) throw err;
